Setup razor on currently running alpine instance
a more modular script using bash functions to setup razor and then generate
a initramfs that is PXE bootable. the goal for this script is to
generate a tarball with vmlinuz and pxe-initramfs which can be placed on
razor server. does not create a iso because everything is included in
the PXE bootable initramfs