start_pre() installs apks and verifies scripts
hopefully the service will not have to do this and everything is already
shoved into our pxe-initramfs but just to be sure this service can start
correctly, install facter and razor-mk-agent and then make sure all the
scripts exist. go to setup-razor-env.sh and look at
verify_pxe_initramfs() for more info.