More TODOs, mkinitfs and abuild-keygen added
Added notes and todo comments about future of this script. We could
actually put everything we need into a apkovl.tar.gz and just use that
tarball. this would also allow us to use Alpine for x86 as well, and
just make sure the tarball works on any architecture (if we want to).